Create homepage and homepage specific css file
Previously, the homepage only contained a few links to the recipes. It
has since been changed to a homepage that includes the website logo,
slogan, and description. Other features like a quote and contact section
must be added to the homepage. Furthermore, the recipe pages must also
be styled to fit the new colour scheme.